Fri, Mar 22, 2024, 12:40 PM -03:00

+++ readability/readability.c
@@ -54,7 +54,7 @@ int test(string text)
             L++;
         }
         tL = (L*100)/nw;
-        tsentences = (sentences*100)/nw;
+        tsentences = (sentences*100)/(float)nw;
     }
     float index = 0.0588 * tL - 0.296 * tsentences - 15.8;
     return (round(index));
